+++ src/Argument/Option.php
@@ -7,43 +7,34 @@
 
 declare(strict_types=1);
 
-namespace DecodeLabs\Commandment\Request\Argument;
+namespace DecodeLabs\Commandment\Argument;
 
-use DecodeLabs\Commandment\Request\Argument;
-use DecodeLabs\Commandment\Request\ArgumentTrait;
+use Attribute;
+use DecodeLabs\Commandment\Argument;
+use DecodeLabs\Commandment\ArgumentTrait;
 
-class ValueList implements Argument
+#[Attribute(Attribute::TARGET_CLASS | Attribute::IS_REPEATABLE)]
+class Option implements Argument
 {
     use ArgumentTrait;
     use ValueTrait;
 
-    /**
-     * @var ?list<string>
-     */
-    protected(set) ?array $default = null;
-
-    public ?int $min = null;
-    public ?int $max = null;
+    protected(set) ?string $default = null;
 
     /**
-     * @param ?list<string> $default
      * @param ?list<string> $options
      */
     public function __construct(
         string $name,
         bool $required = false,
-        ?array $default = null,
+        ?string $default = null,
         ?array $options = null,
-        ?int $min = null,
-        ?int $max = null,
         ?string $description = null
     ) {
         $this->name = $name;
         $this->required = $required;
         $this->default = $default;
         $this->options = $options;
-        $this->min = $min;
-        $this->max = $max;
         $this->description = $description;
     }
 }